revise search button function and move function to check update
