+++ screens/main/checkUpdate.js
